+++ Python/gc_free_threading.c
@@ -11,7 +11,7 @@
 #include "pycore_object_stack.h"
 #include "pycore_pyerrors.h"
 #include "pycore_pystate.h"       // _PyThreadState_GET()
-#include "pycore_tagged.h"        // _PyObject_SetDeferredRefcount
+#include "pycore_stackref.h"        // _PyObject_SetDeferredRefcount
 #include "pycore_time.h"          // _PyTime_GetPerfCounter()
 #include "pycore_tstate.h"        // _PyThreadStateImpl
 #include "pycore_weakref.h"       // _PyWeakref_ClearRef()
